Refactor : PlaceResponseDto 리팩토링
