+++ src/main/java/com/example/malang/dto/PostRequestDto.java
@@ -1,0 +1,4 @@
+package com.example.malang.dto;
+
+public class PostRequestDto {
+}
