+++ src/main/resources/application.yml
@@ -2,39 +2,7 @@ spring:
   main:
     allow-bean-definition-overriding: true
     allow-circular-references: true
-  security:
-    oauth2:
-      client:
-        registration:
-          naver:
-            client-id: ${NAVER_CLIENT_ID}
-            client-secret: ${NAVER_CLIENT_SECRET}
-            authorization-grant-type: authorization_code
-            client-name: inu-malang
-            redirect-uri: http://localhost:8080/login/oauth2/code/naver
-            scope: profile,email
 
-          kakao:
-            client-id: ${KAKAO_CLIENT_ID}
-            client-secret: ${KAKAO_CLIENT_SECRET}
-            authorization-grant-type: authorization_code
-            client-name: inu-malang
-            redirect-uri: http://localhost:8080/login/oauth2/code/kakao
-            scope: profile_nickname,profile_image
-
-        provider:
-          naver:
-            authorization-uri: https://nid.naver.com/oauth2.0/authorize
-            token-uri: https://nid.naver.com/oauth2.0/token
-            user-info-uri: https://openapi.naver.com/v1/nid/me
-            user-name-attribute: response
-          kakao:
-            issuer-uri: https://kauth.kakao.com
-            authorization-uri: https://kauth.kakao.com/oauth/authorize
-            token-uri: https://kauth.kakao.com/oauth/token
-            user-info-uri: https://kapi.kakao.com/v2/user/me
-            jwk-set-uri: https://kauth.kakao.com/.well-known/jwks.json
-            user-name-attribute: id
   profiles:
     active: local
   datasource:
